refactor: move get_community_by_username to orm module and update session type annotation
